requirements.txt
autopep8==1.7.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.9.1
python-dotenv==0.20.0
rsa==4.9
Telethon==1.24.0
toml==0.10.2
